drf_project/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.2
pytest-django==4.5.2
pytest==6.2.5
gunicorn==20.1.0
pytest-cov==3.0.0
flake8==4.0.1
black==22.3.0
whitenoise==5.3.0
click==8.0.2
isort==5.10.1